backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.txt
#
aniso8601==7.0.0
    # via
    #   -r requirements.txt
    #   graphene
asgiref==3.4.1
    # via
    #   -r requirements.txt
    #   django
certifi==2019.9.11
    # via
    #   -r requirements.txt
    #   pipenv
click==8.0.1
    # via
    #   -r requirements.txt
    #   pip-tools
django==3.2.5
    # via
    #   -r requirements.txt
    #   django-cors-headers
    #   django-filter
    #   django-graphql-auth
    #   django-graphql-jwt
    #   graphene-django
django-cors-headers==3.7.0
    # via -r requirements.txt
django-filter==2.4.0
    # via
    #   -r requirements.txt
    #   django-graphql-auth
django-graphql-auth==0.3.16
    # via -r requirements.txt
django-graphql-jwt==0.3.0
    # via
    #   -r requirements.txt
    #   django-graphql-auth
graphene==2.1.8
    # via
    #   -r requirements.txt
    #   django-graphql-auth
    #   graphene-django
graphene-django==2.15.0
    # via
    #   -r requirements.txt
    #   django-graphql-auth
    #   django-graphql-jwt
graphql-core==2.3.2
    # via
    #   -r requirements.txt
    #   django-graphql-jwt
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via
    #   -r requirements.txt
    #   graphene
pep517==0.10.0
    # via
    #   -r requirements.txt
    #   pip-tools
pip-tools==6.2.0
    # via -r requirements.txt
pipenv==2018.11.26
    # via -r requirements.txt
promise==2.3
    # via
    #   -r requirements.txt
    #   graphene-django
    #   graphql-core
    #   graphql-relay
psycopg2-binary==2.9.1
    # via -r requirements.txt
pyjwt==1.7.1
    # via
    #   -r requirements.txt
    #   django-graphql-auth
    #   django-graphql-jwt
pytz==2021.1
    # via
    #   -r requirements.txt
    #   django
rx==1.6.1
    # via
    #   -r requirements.txt
    #   graphql-core
singledispatch==3.6.2
    # via
    #   -r requirements.txt
    #   graphene-django
six==1.16.0
    # via
    #   -r requirements.txt
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   promise
    #   singledispatch
sqlparse==0.4.1
    # via
    #   -r requirements.txt
    #   django
text-unidecode==1.3
    # via
    #   -r requirements.txt
    #   graphene-django
toml==0.10.2
    # via
    #   -r requirements.txt
    #   pep517
virtualenv==16.7.7
    # via
    #   -r requirements.txt
    #   pipenv
virtualenv-clone==0.5.3
    # via
    #   -r requirements.txt
    #   pipenv
wheel==0.36.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
